+++ tests/derive/enum.rs
@@ -66,3 +66,41 @@ fn not_constructible() {
     assert!(sscanf!("5", "{Number}").is_err());
     assert!(sscanf!("-1/2", "{Number}").is_err());
 }
+
+#[test]
+fn transparent() {
+    #[derive(FromScanf, Debug, PartialEq)]
+    enum Dynamic {
+        #[sscanf(transparent)]
+        Number(isize),
+        #[sscanf(transparent)]
+        String(String),
+    }
+
+    let input = "123 hello 0 hi";
+    let (num, string, zero, hi) =
+        sscanf!(input, "{Dynamic} {Dynamic} {Dynamic} {Dynamic}").unwrap();
+    assert_eq!(num, Dynamic::Number(123));
+    assert_eq!(string, Dynamic::String("hello".to_string()));
+    assert_eq!(zero, Dynamic::Number(0));
+    assert_eq!(hi, Dynamic::String("hi".to_string()));
+}
+
+#[test]
+fn autogen() {
+    #[derive(FromScanf, Debug, PartialEq)]
+    #[sscanf(autogen = "CaseSensitive")]
+    enum Words {
+        Hello,
+        World,
+        Hi,
+    }
+
+    let input = "Hello World Hi";
+    let expected = (Words::Hello, Words::World, Words::Hi);
+    let parsed = sscanf!(input, "{Words} {Words} {Words}").unwrap();
+    assert_eq!(parsed, expected);
+
+    let input_lower = "hello world hi";
+    sscanf!(input_lower, "{Words} {Words} {Words}").unwrap_err();
+}
