+++ tests/fail/nightly/input_not_str.stderr
@@ -4,13 +4,13 @@ error[E0308]: mismatched types
 2 |     sscanf::sscanf!(5usize, "{}", usize);
   |     ----------------^^^^^^--------------
   |     |               |
-  |     |               expected `str`, found `usize`
+  |     |               expected `&str`, found `&usize`
   |     arguments to this function are incorrect
   |
   = note: expected reference `&str`
              found reference `&usize`
-note: associated function defined here
- --> /rustc/52372f9c71d8ade4cb815524f179119656f0aa2e/library/core/src/str/mod.rs:441:18
+note: method defined here
+ --> $RUST/core/src/str/mod.rs
 
 error[E0308]: mismatched types
  --> tests/fail/nightly/input_not_str.rs:3:21
@@ -18,10 +18,10 @@ error[E0308]: mismatched types
 3 |     sscanf::sscanf!(b"5", "{}", usize);
   |     ----------------^^^^--------------
   |     |               |
-  |     |               expected `str`, found `&[u8; 1]`
+  |     |               expected `&str`, found `&&[u8; 1]`
   |     arguments to this function are incorrect
   |
   = note: expected reference `&str`
              found reference `&&'static [u8; 1]`
-note: associated function defined here
- --> /rustc/52372f9c71d8ade4cb815524f179119656f0aa2e/library/core/src/str/mod.rs:441:18
+note: method defined here
+ --> $RUST/core/src/str/mod.rs
