+++ tests/fail/nightly/str_lifetime.rs
