+++ tests/fail/nightly/str_lifetime.stderr
@@ -1,0 +1,23 @@
+error[E0597]: `input` does not live long enough
+ --> tests/fail/nightly/str_lifetime.rs:5:29
+  |
+4 |         let input = String::from("hi");
+  |             ----- binding `input` declared here
+5 |         s = sscanf::sscanf!(input, "{str}").unwrap();
+  |                             ^^^^^ borrowed value does not live long enough
+6 |     }
+  |     - `input` dropped here while still borrowed
+7 |     println!("{}", s);
+  |                    - borrow later used here
+
+error[E0597]: `input` does not live long enough
+  --> tests/fail/nightly/str_lifetime.rs:16:29
+   |
+15 |         let input = String::from("hi");
+   |             ----- binding `input` declared here
+16 |         w = sscanf::sscanf!(input, "{Wrapper}").unwrap();
+   |                             ^^^^^ borrowed value does not live long enough
+17 |     }
+   |     - `input` dropped here while still borrowed
+18 |     println!("{}", w.0);
+   |                    --- borrow later used here
