+++ tests/fail/stable/input_not_str.stderr
@@ -10,6 +10,7 @@ error[E0308]: mismatched types
   = note: expected reference `&str`
              found reference `&usize`
 note: associated function defined here
+ --> $RUST/core/src/str/mod.rs
 
 error[E0308]: mismatched types
  --> tests/fail/stable/input_not_str.rs:3:21
@@ -23,3 +24,4 @@ error[E0308]: mismatched types
   = note: expected reference `&str`
              found reference `&&'static [u8; 1]`
 note: associated function defined here
+ --> $RUST/core/src/str/mod.rs
