+++ tests/fail/stable/invalid_type_in_placeholder.stderr
@@ -38,18 +38,34 @@ error[E0412]: cannot find type `bob` in this scope
 5 |     sscanf::sscanf!("hi", "{bob:/hi/}");
   |                           ^^^^^^^^^^^^ not found in this scope
 
-error[E0080]: evaluation of constant value failed
+note: erroneous constant used
  --> tests/fail/stable/invalid_type_in_placeholder.rs:3:5
   |
 3 |     sscanf::sscanf!("hi", "{bob}");
-  |     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ referenced constant has errors
+  |     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
   |
-  = note: this error originates in the macro `$crate::pmr::__concatcp_impl` which comes from the expansion of the macro `sscanf::sscanf` (in Nightly builds, run with -Z macro-backtrace for more info)
+  = note: this note originates in the macro `$crate::pmr::__concatcp_impl` which comes from the expansion of the macro `sscanf::sscanf` (in Nightly builds, run with -Z macro-backtrace for more info)
 
-error[E0080]: evaluation of constant value failed
+note: erroneous constant used
  --> tests/fail/stable/invalid_type_in_placeholder.rs:3:5
   |
 3 |     sscanf::sscanf!("hi", "{bob}");
-  |     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ referenced constant has errors
+  |     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
   |
-  = note: this error originates in the macro `__cf_osRcTFl4A::__concatcp_inner` which comes from the expansion of the macro `sscanf::sscanf` (in Nightly builds, run with -Z macro-backtrace for more info)
+  = note: this note originates in the macro `__cf_osRcTFl4A::__concatcp_inner` which comes from the expansion of the macro `sscanf::sscanf` (in Nightly builds, run with -Z macro-backtrace for more info)
+
+note: erroneous constant used
+ --> tests/fail/stable/invalid_type_in_placeholder.rs:3:5
+  |
+3 |     sscanf::sscanf!("hi", "{bob}");
+  |     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+  |
+  = note: this note originates in the macro `__cf_osRcTFl4A::__concatcp_inner` which comes from the expansion of the macro `sscanf::sscanf` (in Nightly builds, run with -Z macro-backtrace for more info)
+
+note: erroneous constant used
+ --> tests/fail/stable/invalid_type_in_placeholder.rs:3:5
+  |
+3 |     sscanf::sscanf!("hi", "{bob}");
+  |     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+  |
+  = note: this note originates in the macro `__cf_osRcTFl4A::__concatcp_inner` which comes from the expansion of the macro `sscanf::sscanf` (in Nightly builds, run with -Z macro-backtrace for more info)
