+++ tests/fail/stable/str_lifetime.stderr
@@ -1,5 +1,5 @@
 error[E0597]: `input` does not live long enough
- --> tests/fail/str_lifetime.rs:5:29
+ --> tests/fail/stable/str_lifetime.rs:5:29
   |
 5 |         s = sscanf::sscanf!(input, "{str}").unwrap();
   |                             ^^^^^ borrowed value does not live long enough
@@ -9,7 +9,7 @@ error[E0597]: `input` does not live long enough
   |                    - borrow later used here
 
 error[E0597]: `input` does not live long enough
-  --> tests/fail/str_lifetime.rs:16:29
+  --> tests/fail/stable/str_lifetime.rs:16:29
    |
 16 |         w = sscanf::sscanf!(input, "{Wrapper}").unwrap();
    |                             ^^^^^ borrowed value does not live long enough
